[FIX] Fixes error with unsigned parralel for variable in vs.

+++ apps/gustaf/stellar_routines.h
@@ -286,7 +286,7 @@ void _getMatchDistanceScore(
     Splitter<TQueryMatchSetIterator> setSplitter(begin(queryMatchesSet, Standard()), end(queryMatchesSet, Standard()));
 
     SEQAN_OMP_PRAGMA(parallel for)
-    for(unsigned jobId = 0; jobId < length(setSplitter); ++jobId)
+    for(int jobId = 0; jobId < static_cast<int>(length(setSplitter)); ++jobId)
     {
         for (TQueryMatchSetIterator it = setSplitter[jobId]; it != setSplitter[jobId + 1]; ++it)
         {
